Move Section out of Search
