+++ plugin-hrm-form/src/components/Section.jsx
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types';
 import { ButtonBase, Collapse } from '@material-ui/core';
 import { ArrowDropDownTwoTone, ArrowDropUpTwoTone } from '@material-ui/icons';
 
-import { SectionTitleContainer, SectionTitleText, ContactDetailsIcon } from '../../../styles/search';
+import { SectionTitleContainer, SectionTitleText, ContactDetailsIcon } from '../styles/search';
 
 const ArrowDownIcon = ContactDetailsIcon(ArrowDropDownTwoTone);
 const ArrowUpIcon = ContactDetailsIcon(ArrowDropUpTwoTone);
