+++ plugin-hrm-form/src/components/search/ContactDetails/Details.jsx
@@ -5,7 +5,7 @@ import { ButtonBase, IconButton } from '@material-ui/core';
 import { MoreHoriz, Link as LinkIcon } from '@material-ui/icons';
 
 import { DetailsContainer, NameContainer, DetNameText, ContactDetailsIcon } from '../../../styles/search';
-import Section from './Section';
+import Section from '../../Section';
 import SectionEntry from './SectionEntry';
 import callTypes, { channelTypes } from '../../../states/DomainConstants';
 import { isNonDataCallType } from '../../../states/ValidationRules';
